Remove unnecessary i32 cast in sys_getpid call
Fix clippy error about unnecessary type casting. The sys_getpid() method
already returns i32, so casting it to i32 is redundant and triggers a
clippy::unnecessary_cast lint error in CI.

+++ litebox_shim_linux/src/syscalls/file.rs
@@ -1280,7 +1280,7 @@ impl Task {
                 // In LiteBox, processes are simplified: each process forms its own process group
                 // where the process group ID equals the process ID. This is a simplification
                 // compared to standard Unix where processes can explicitly join different groups.
-                let pid = self.sys_getpid() as i32;
+                let pid = self.sys_getpid();
                 pgrp.write_at_offset(0, pid).ok_or(Errno::EFAULT)?;
                 Ok(0)
             }
